+++ ruizsalas/src/main/java/basededatos/iUsuario_No_Registrado.java
@@ -7,11 +7,11 @@ public interface iUsuario_No_Registrado extends iUsuario {
 
 	public void enviar_Correo_Confirmacion(String aApodo, String aDni, String aCorreo);
 
-	public void gestionar_Transaccion(String aApodo, String aDni, String aCorreo, int aTarjeta_de_credito);
+	public void gestionar_Transaccion(String aApodo, String aDni, String aCorreo, long aTarjeta_de_credito);
 
 	public void enviar_Correo_Confirmacion();
 
-	public Usuario_suscrito_ gestionar_Transaccion(String aCorreo, String aContrasena, String aApodo, String aFoto_de_perfil, String aDni, int aTarjeta_de_credito);
+	public Usuario_suscrito_ gestionar_Transaccion(String aCorreo, String aContrasena, String aApodo, String aFoto_de_perfil, String aDni, long aTarjeta_de_credito);
 
 	public Usuario confirmar_Datos(String aCorreo, String aContrasena);
 }
