+++ ruizsalas/src/main/java/basededatos/iUsuario_Registardo.java
@@ -2,7 +2,7 @@ package basededatos;
 
 public interface iUsuario_Registardo extends iRegistrado {
 
-	public void cancelar_cobro_suscripcion(String aApodo, String aDni, String aCorreo, int aTarjeta_de_credito);
+	public void cancelar_cobro_suscripcion(String aApodo, String aDni, String aCorreo, long aTarjeta_de_credito);
 
 	public void cancelar_cobro_suscripcion(int aIdUsuario);
 }
